+++ src/xyz/oihana/schema/constants/traits/statistics/CustomerStatisticsTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\statistics\CustomerStatistics} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.4.0
+ */
+trait CustomerStatisticsTrait
+{
+    const string ASSIGNED_POS    = 'assignedPOS'    ;
+    const string ASSIGNED_SELLER = 'assignedSeller' ;
+}