+++ src/xyz/oihana/schema/constants/traits/statistics/HasTradingMeasuresTrait.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\traits\HasTradingMeasures} trait.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.4.0
+ */
+trait HasTradingMeasuresTrait
+{
+    const string AVERAGE_COST    = 'averageCost'    ;
+    const string AVERAGE_MARGIN  = 'averageMargin'  ;
+    const string COST_PRICE      = 'costPrice'      ;
+    const string GROSS_MARGIN    = 'grossMargin'    ;
+    const string PURCHASE_COST   = 'purchaseCost'   ;
+    const string PURCHASE_MARGIN = 'purchaseMargin' ;
+    const string QUANTITY        = 'quantity'       ;
+    const string REVENUE         = 'revenue'        ;
+    const string VOLUME          = 'volume'         ;
+    const string WEIGHT          = 'weight'         ;
+}